Reformatted code
Code was reformatted

The api endpoint has been changed to api/weatherdata
According to the REST API guidelines you should not put POST/GET/PUT/DELETE in the name, the endpoint should figure it out itself.
Getting the latest station data is integrated in this endpoint.

Also put code in folders to make it more readable, changed some function names to make it more readable.

Changed to tsconfig file to allow for better importing, you can now import hooks for example by using the following format: @hooks/[filename].

This is also done with types, but @types is reserved for types found in node_modules thus to import our own types you must prefix it with a 'c' eg. @ctypes[filename]

Also made a custom hook to get the station data.
Added an example for this hook in pages/station/[id].tsx

Furthermore added a new data type called StationData; before we used WeatherData as a return type for our GET api request.
This datatype was meant for the POST request because the delivered data is different then how we store our data in the database.
See differences in @ctypes/types